FirebaseBaseTimestampSerializer introduced
